timely_django/requirements.txt
asgiref==3.3.3
certifi==2020.12.5
chardet==4.0.0
Django==3.2
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
django-dynamic-formsets==0.0.8
django-rest-allauth==1.1
django-rest-multiple-models==2.1.3
djangorestframework==3.12.4
idna==2.10
numpy==1.20.2
pandas==1.2.3
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
